python/requirements.txt
# Flask Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# HTTP Client
httpx>=0.27.0

# Environment Variables
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Type Hints
typing-extensions>=4.9.0
